pyproject.toml
[project]
name = "estudia-mcp"
version = "0.1.0"
description = "MCP Server for RAG with Supabase and Gemini"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.13.0.2",
    "google-generativeai>=0.3.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
]
